+++ src/shared/replication/rules.rs
@@ -21,7 +21,7 @@ pub trait AppRuleExt {
     ///
     /// If present on an entity with [`Replicated`] component,
     /// it will be serialized and deserialized as-is using [`postcard`]
-    /// and sent at [`SendRate::EveryTick`]. To customize this, use [`Self::replicate_with`].
+    /// and sent at [`ReplicationMode::OnChange`]. To customize this, use [`Self::replicate_with`].
     ///
     /// See also the section on [`components`](../../index.html#components) from the quick start guide.
     fn replicate<C>(&mut self) -> &mut Self
@@ -31,7 +31,7 @@ pub trait AppRuleExt {
         self.replicate_filtered::<C, ()>()
     }
 
-    /// Like [`Self::replicate`], but uses [`SendRate::Once`].
+    /// Like [`Self::replicate`], but uses [`ReplicationMode::Once`].
     fn replicate_once<C>(&mut self) -> &mut Self
     where
         C: Component<Mutability: MutWrite<C>> + Serialize + DeserializeOwned,
@@ -79,7 +79,7 @@ pub trait AppRuleExt {
     where
         C: Component<Mutability: MutWrite<C>> + Serialize + DeserializeOwned,
     {
-        self.replicate_with_filtered::<_, F>((RuleFns::<C>::default(), SendRate::Once))
+        self.replicate_with_filtered::<_, F>((RuleFns::<C>::default(), ReplicationMode::Once))
     }
 
     /**
@@ -90,7 +90,7 @@ pub trait AppRuleExt {
 
     You can also pass a tuple of [`RuleFns`] to define a rule for multiple components.
     These components will only be replicated if all of them are present on the entity.
-    To assign a [`SendRate`] to a component, wrap its [`RuleFns`] in a tuple with the
+    To assign a [`ReplicationMode`] to a component, wrap its [`RuleFns`] in a tuple with the
     desired rate.
 
     If an entity matches multiple rules, the functions from the rule with higher priority
@@ -192,7 +192,7 @@ pub trait AppRuleExt {
     .replicate_with((
         RuleFns::<MovingPlatform>::default(),
         // Send position only once.
-        (RuleFns::<Position>::default(), SendRate::Once),
+        (RuleFns::<Position>::default(), ReplicationMode::Once),
     ));
 
     #[derive(Component, Deserialize, Serialize)]
@@ -395,7 +395,7 @@ pub trait AppRuleExt {
     /// # app.add_plugins(RepliconPlugins);
     /// app.replicate_with_filtered::<_, With<StaticBox>>((
     ///     RuleFns::<Health>::default(),
-    ///     (RuleFns::<Transform>::default(), SendRate::Once),
+    ///     (RuleFns::<Transform>::default(), ReplicationMode::Once),
     /// ));
     /// # #[derive(Component)]
     /// # struct StaticBox;
@@ -436,7 +436,7 @@ pub trait AppRuleExt {
     Defines a [`ReplicationRule`] for a bundle.
 
     Implemented for tuples of components. Use it to conveniently create a rule with
-    default ser/de functions and [`SendRate::EveryTick`] for all components.
+    default ser/de functions and [`ReplicationMode::OnChange`] for all components.
     To customize this, use [`Self::replicate_with`].
 
     Can also be implemented manually for user-defined bundles.
@@ -736,13 +736,13 @@ mod tests {
         app.init_resource::<ProtocolHasher>()
             .init_resource::<ReplicationRules>()
             .init_resource::<ReplicationRegistry>()
-            .replicate_with((RuleFns::<A>::default(), SendRate::Once))
+            .replicate_with((RuleFns::<A>::default(), ReplicationMode::Once))
             .replicate_with((RuleFns::<B>::default(), RuleFns::<C>::default()))
             .replicate_with_priority(
                 4,
                 (
                     RuleFns::<C>::default(),
-                    (RuleFns::<D>::default(), SendRate::Once),
+                    (RuleFns::<D>::default(), ReplicationMode::Once),
                 ),
             );
 
@@ -782,7 +782,7 @@ mod tests {
         app.init_resource::<ProtocolHasher>()
             .init_resource::<ReplicationRules>()
             .init_resource::<ReplicationRegistry>()
-            .replicate_with_filtered::<_, With<B>>((RuleFns::<A>::default(), SendRate::Once))
+            .replicate_with_filtered::<_, With<B>>((RuleFns::<A>::default(), ReplicationMode::Once))
             .replicate_with_filtered::<_, Or<(With<A>, With<D>)>>((
                 RuleFns::<B>::default(),
                 RuleFns::<C>::default(),
@@ -791,7 +791,7 @@ mod tests {
                 5,
                 (
                     RuleFns::<C>::default(),
-                    (RuleFns::<D>::default(), SendRate::Once),
+                    (RuleFns::<D>::default(), ReplicationMode::Once),
                 ),
             );